+++ src/pizzatales/RifleBullet.java
@@ -6,7 +6,7 @@ public class RifleBullet extends Projectile {
 	
 	public static Image bulletsprite;
 	
-	public RifleBullet(int startX, int startY, int vectorX, int vectorY, int speed, float dmg, int range) {
+	public RifleBullet(int startX, int startY, float vectorX, float vectorY, int speed, float dmg, int range) {
 		super(startX, startY, vectorX, vectorY, speed, dmg, 15, 7, range);
 	}
