+++ src/pizzatales/SmgBullet.java
@@ -6,7 +6,7 @@ public class SmgBullet extends Projectile {
 	
 	public static Image bulletsprite;
 	
-	public SmgBullet(int startX, int startY, int vectorX, int vectorY, int speed, float dmg, int range) {
+	public SmgBullet(int startX, int startY, float vectorX, float vectorY, int speed, float dmg, int range) {
 		super(startX, startY, vectorX, vectorY, speed, dmg, 15, 7, range);
 	}
